app/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
psycopg2==2.9.3
PyJWT==2.3.0
python-dotenv==0.18.0